Add get_probs function for PROLBL

+++ inc/NuFastLBL.h
@@ -20,9 +20,9 @@ namespace NuFastLBL {
 //   N_Newton: number of Newton's method iterations to do. should be zero, one, two (or higher)
 // Outputs:
 //   probs_returned is all nine oscillation probabilities: e.g. probs_returned[1][0] is mu->e
-void Probability_Matter_LBL(double s12sq, double s13sq, double s23sq, double delta, double Dmsq21, double Dmsq31, double L, double E, double rho, double Ye, int N_Newton, double (*probs_returned)[3][3]);
+void Probability_Matter_LBL(double s12sq, double s13sq, double s23sq, double delta, double Dmsq21, double Dmsq31, double L, double E, double rho, double Ye, int N_Newton, float (*probs_returned)[3][3]);
 
-void Probability_Vacuum_LBL(double s12sq, double s13sq, double s23sq, double delta, double Dmsq21, double Dmsq31, double L, double E, double (*probs_returned)[3][3]);
+void Probability_Vacuum_LBL(double s12sq, double s13sq, double s23sq, double delta, double Dmsq21, double Dmsq31, double L, double E, float (*probs_returned)[3][3]);
 
 }
