+++ src/NuFastLBL.cxx
@@ -1,4 +1,5 @@
 // This code was taken from NuFast-LBL and was originally licensed under the license below
+// The probs_returned parameters are changed from double to float to match PROfit conventions.
 // https://github.com/PeterDenton/NuFast-LBL
 // arXiv:2405.02400
 
@@ -52,7 +53,7 @@ constexpr double const YerhoE2a = 1.52588e-4;
 //   N_Newton: number of Newton's method iterations to do. should be zero, one, two (or higher)
 // Outputs:
 //   probs_returned is all nine oscillation probabilities: e.g. probs_returned[1][0] is mu->e
-void Probability_Matter_LBL(double s12sq, double s13sq, double s23sq, double delta, double Dmsq21, double Dmsq31, double L, double E, double rho, double Ye, int N_Newton, double (*probs_returned)[3][3])
+void Probability_Matter_LBL(double s12sq, double s13sq, double s23sq, double delta, double Dmsq21, double Dmsq31, double L, double E, double rho, double Ye, int N_Newton, float (*probs_returned)[3][3])
 {
 	double c13sq, sind, cosd, Jrr, Jmatter, Dmsqee, Amatter;
 	double Ue1sq, Ue2sq, Ue3sq, Um1sq, Um2sq, Um3sq, Ut1sq, Ut2sq, Ut3sq;
@@ -203,7 +204,7 @@ void Probability_Matter_LBL(double s12sq, double s13sq, double s23sq, double del
 	(*probs_returned)[2][2] = 1 - (*probs_returned)[0][2] - (*probs_returned)[1][2];	// Ptt
 }
 
-void Probability_Vacuum_LBL(double s12sq, double s13sq, double s23sq, double delta, double Dmsq21, double Dmsq31, double L, double E, double (*probs_returned)[3][3])
+void Probability_Vacuum_LBL(double s12sq, double s13sq, double s23sq, double delta, double Dmsq21, double Dmsq31, double L, double E, float (*probs_returned)[3][3])
 {
 	double c13sq, sind, cosd, Jrr, Jvac;
 	double Ue1sq, Ue2sq, Ue3sq, Um1sq, Um2sq, Um3sq, Ut1sq, Ut2sq, Ut3sq;
